attach pkg:/system/dtrace/tests to the build
Primarily so that we may use `dstyle` as part of `make check`, at the moment.

+++ usr/src/cmd/dtrace/test/tst/common/ustack/tst.spin.c
@@ -55,5 +55,5 @@ foo(int a, int b)
 int
 main(int argc, char **argv)
 {
-	return (foo(argc, (int)argv) == 0);
+	return (foo(argc, (int)(intptr_t)argv) == 0);
 }
